+++ src/prices/Types.ts
@@ -7,9 +7,6 @@ export type PriceSourcesConfig = {
     /** Endpoints for backend price data */
     backendEndpoints: string[];
     
-    /** Endpoints for IOTA price data */
-    iotaEndpoints: string[];
-    
     /** Endpoints for ICP price data */
     icpEndpoints: string[];
 };
@@ -18,8 +15,7 @@ export type PriceSourcesConfig = {
  * Default configuration for price source endpoints.
  */
 export const DefaultPriceSourcesConfig: PriceSourcesConfig = {
-    backendEndpoints: ['evaa.space'],
-    iotaEndpoints: ['api.stardust-mainnet.iotaledger.net'],
+    backendEndpoints: ['api.evaa.space', 'evaa.space'],
     icpEndpoints: ['6khmc-aiaaa-aaaap-ansfq-cai.raw.icp0.io'],
 }