+++ src/utils/userJettonWallet.ts
@@ -22,6 +22,12 @@ function getUserJettonData(ownerAddress: Address, assetName: string, jettonWalle
               .storeAddress(ownerAddress)
               .storeAddress(jettonMasterAddress)
               .endCell();
+      case 'PT_tsUSDe_01Sep2025':
+        return beginCell()
+              .storeCoins(0)
+              .storeAddress(ownerAddress)
+              .storeAddress(jettonMasterAddress)
+              .endCell();
       case 'tsUSDe':
         return beginCell()
           .storeUint(0, 4)